+++ src/main/java/com/czy/template/pojo/Exam.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class Exam {
     private int exam_id;                 //考试id
-    private int create_id;               //创建教师id
+    private Long create_id;               //创建教师id
     private String title;                 //考试名称
     private String status;               //考试状态
     private String questionIds;    //考试题目
