+++ src/main/java/com/czy/template/pojo/User.java
@@ -1,7 +1,7 @@
 package com.czy.template.pojo;
 
 public class User {
-    private int id;
+    private Long id;
     private String realname;
     private String username;
     private String password;
@@ -11,7 +11,7 @@ public class User {
     private String  address;
     private int identity;
 
-    public User(int id, String realname, String username, String password, String phone, String email, char gender, String address, int identity) {
+    public User(Long id, String realname, String username, String password, String phone, String email, char gender, String address, int identity) {
         this.id = id;
         this.realname = realname;
         this.username = username;
@@ -26,11 +26,11 @@ public class User {
     public User() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
